refactor(collector): fix demultiplexer imports, move NoneModule to noop-impl, inline test helpers
